move private headers and package sdk in single folder (#28)
## What does this PR do ?
Moved internal headers in an internel subfolder.
keep only kuzzle.hpp as public header.

Now, the user only need to include   `kuzzle.hpp` to use Kuzzle and its controlers

### How should this be manually tested?
Build the sdk: `make`
Build the packeage `make package`
Extract the packege, and try to build an app using it
