+++ Makefile
@@ -24,42 +24,42 @@ else
 	MV = mv -f
 	ROOT_DIR_CLEAN = $(ROOT_DIR)
 	LIB_PREFIX = lib
+	ARCH=$(shell uname -p)
 endif
 
+SDK_FOLDER_NAME=kuzzle-cpp-sdk
+
+export GOPATH = $(ROOT_DIR)go
+
 PATHSEP = $(strip $(SEP))
-ROOTOUTDIR = $(ROOT_DIR)build
-CPPDIR = $(ROOT_DIR)cpp
-HEADERSDIR = $(ROOT_DIR)include
+BUILD_DIR = build
 AR ?= ar
 
-CXXFLAGS = -g -fPIC -std=c++11 -I.$(PATHSEP)include -I.$(PATHSEP)sdk-c$(PATHSEP)include$(PATHSEP) -I.$(PATHSEP)sdk-c$(PATHSEP)build$(PATHSEP) -L.$(PATHSEP)sdk-c$(PATHSEP)build
+CXXFLAGS = -g -fPIC -std=c++11 -MMD \
+	-I.$(PATHSEP)include \
+	-I.$(PATHSEP)sdk-c$(PATHSEP)build$(PATHSEP)kuzzle-c-sdk$(PATHSEP)include \
+	-L.$(PATHSEP)sdk-c$(PATHSEP)build$(PATHSEP)kuzzle-c-sdk$(PATHSEP)lib
+
 LDFLAGS = -lkuzzlesdk
 
-CPP_SDK_SRCS = src$(PATHSEP)kuzzle.cpp \
-					src$(PATHSEP)auth.cpp \
-					src$(PATHSEP)collection.cpp \
-					src$(PATHSEP)document.cpp \
-					src$(PATHSEP)index.cpp \
-					src$(PATHSEP)realtime.cpp \
-					src$(PATHSEP)search_result.cpp \
-					src$(PATHSEP)server.cpp \
-					src$(PATHSEP)index.cpp \
-					src$(PATHSEP)websocket.cpp \
-					src$(PATHSEP)default_constructors.cpp \
-					src$(PATHSEP)specification_search_result.cpp
+SOURCES = $(wildcard src$(PATHSEP)*.cpp)
 
-CPPSDK = $(CPP_SDK_SRCS:%.cpp=%.o)
+OBJECTS = $(SOURCES:%.cpp=$(BUILD_DIR)/%.o)
+DEPENDS = $(OBJECTS:%.o=%.d)
 
-all: cpp
+all: $(BUILD_DIR)/cpp
 
-%.o: %.cpp
+-include ${DEPENDS}
+
+$(BUILD_DIR)/%.o: %.cpp
+	mkdir -p $(@D)
 	$(CXX) -fPIC -c $< -o $@ $(CXXFLAGS) $(LDFLAGS)
 
-makedir:
+$(BUILD_DIR):
 ifeq ($(OS),Windows_NT)
-	@if not exist $(subst /,\,$(ROOTOUTDIR)) mkdir $(subst /,\,$(ROOTOUTDIR))
+	@if not exist $(subst /,\,$(BUILD_DIR)) mkdir $(subst /,\,$(BUILD_DIR))
 else
-	mkdir -p $(ROOTOUTDIR)
+	mkdir -p $@
 endif
 
 update_submodule:
@@ -68,45 +68,39 @@ update_submodule:
 make_c_sdk:
 	cd sdk-c && $(MAKE)
 
-cpp: export GOPATH = $(ROOT_DIR)go
-cpp: makedir make_c_sdk $(CPPSDK)
-		$(AR) rvs $(ROOTOUTDIR)$(PATHSEP)libkuzzlesdk$(STATICLIB) src$(PATHSEP)*.o
-		$(CXX) -shared -fPIC -o $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB) -Wl,--whole-archive $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB) sdk-c$(PATHSEP)build$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB) -Wl,--no-whole-archive
-		cd $(ROOTOUTDIR) && mv $(LIB_PREFIX)kuzzlesdk$(STATICLIB) $(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION) && mv $(LIB_PREFIX)kuzzlesdk$(DYNLIB) $(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION)
-		cd $(ROOTOUTDIR) && ln -sr $(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) $(LIB_PREFIX)kuzzlesdk$(DYNLIB)
-		cd $(ROOTOUTDIR) && ln -sr $(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION) $(LIB_PREFIX)kuzzlesdk$(STATICLIB)
+$(BUILD_DIR)/cpp: $(BUILD_DIR) make_c_sdk $(OBJECTS)
+	$(AR) rvs $(BUILD_DIR)$(PATHSEP)libkuzzlesdk$(STATICLIB).$(VERSION) $(OBJECTS)
+	$(CXX) -shared -fPIC -o $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) -Wl,--whole-archive $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION) sdk-c$(PATHSEP)build$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB) -Wl,--no-whole-archive
+	cd $(BUILD_DIR) && ln -srf $(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) $(LIB_PREFIX)kuzzlesdk$(DYNLIB)
+	cd $(BUILD_DIR) && ln -srf $(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION) $(LIB_PREFIX)kuzzlesdk$(STATICLIB)
+	mkdir -p $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/lib
+	mkdir -p $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/include
 
-package: $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION)
-	mkdir $(ROOTOUTDIR)$(PATHSEP)lib
-	mkdir $(ROOTOUTDIR)$(PATHSEP)include
-	cp -fr $(ROOT_DIR)$(PATHSEP)include$(PATHSEP)*.hpp $(ROOTOUTDIR)$(PATHSEP)include
-	cp $(ROOT_DIR)$(PATHSEP)sdk-c$(PATHSEP)include$(PATHSEP)kuzzlesdk.h $(ROOTOUTDIR)$(PATHSEP)include
-	cp $(ROOT_DIR)$(PATHSEP)sdk-c$(PATHSEP)include$(PATHSEP)sdk_wrappers_internal.h $(ROOTOUTDIR)$(PATHSEP)include
-	cp $(ROOT_DIR)$(PATHSEP)sdk-c$(PATHSEP)include$(PATHSEP)protocol.h $(ROOTOUTDIR)$(PATHSEP)include
-	cp $(ROOT_DIR)$(PATHSEP)sdk-c$(PATHSEP)build$(PATHSEP)kuzzle.h $(ROOTOUTDIR)$(PATHSEP)include
-	cp $(ROOTOUTDIR)$(PATHSEP)*.so  $(ROOTOUTDIR)$(PATHSEP)lib
-	cp $(ROOTOUTDIR)$(PATHSEP)*.a  $(ROOTOUTDIR)$(PATHSEP)lib
+	cp -fr include/*  $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/include
+	cp -ra sdk-c$(PATHSEP)build/kuzzle-c-sdk/include/* $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/include
+	cp -a $(BUILD_DIR)$(PATHSEP)*.so*  $(BUILD_DIR)$(PATHSEP)*.a* $(BUILD_DIR)$(PATHSEP)$(SDK_FOLDER_NAME)/lib
+	@touch $@
 
-	mkdir deploy && cd $(ROOTOUTDIR) && tar cfz ..$(PATHSEP)deploy$(PATHSEP)kuzzlesdk-cpp-$(VERSION)-$(ARCH).tar.gz lib include
+package: $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION)
+	mkdir -p deploy && cd $(BUILD_DIR) && tar cfz ..$(PATHSEP)deploy$(PATHSEP)kuzzlesdk-cpp-$(VERSION)-$(ARCH).tar.gz $(SDK_FOLDER_NAME)
 
-build_test: $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) $(ROOTOUTDIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION)
-	cd $(ROOT_DIR)$(PATHSEP)test && sh build_cpp_tests.sh
+build_test: $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(DYNLIB).$(VERSION) $(BUILD_DIR)$(PATHSEP)$(LIB_PREFIX)kuzzlesdk$(STATICLIB).$(VERSION)
+	cd test && sh build_cpp_tests.sh
 
-run_test: $(ROOT_DIR)$(PATHSEP)test$(PATHSEP)_build_cpp_tests$(PATHSEP)KuzzleSDKStepDefs
-	cd $(ROOT_DIR)$(PATHSEP)test && .$(PATHSEP)_build_cpp_tests$(PATHSEP)KuzzleSDKStepDefs > /dev/null &
-	cd $(ROOT_DIR)$(PATHSEP)test && cucumber
+run_test:
+	cd test && .$(PATHSEP)_build_cpp_tests$(PATHSEP)KuzzleSDKStepDefs > /dev/null &
+	cd test && cucumber
 
 clean:
 	cd sdk-c && make clean
 ifeq ($(OS),Windows_NT)
-	$(RRM) $(ROOTOUTDIR)
-	$(RRM) src$(PATHSEP)*.o
+	$(RRM) $(BUILD_DIR)
 	$(RRM) $(ROOT_DIR)$(PATHSEP)deploy
 	$(RRM) $(ROOT_DIR)$(PATHSEP)test$(PATHSEP)_build_cpp_tests
 else
-	$(RRM) $(ROOTOUTDIR) src/*.o $(ROOT_DIR)$(PATHSEP)deploy $(ROOT_DIR)$(PATHSEP)test$(PATHSEP)_build_cpp_tests
+	$(RRM) $(BUILD_DIR) deploy test$(PATHSEP)_build_cpp_tests
 endif
-.PHONY: all cpp core clean
+.PHONY: all core clean run_test 
 
 
 .DEFAULT_GOAL := all
