+++ include/internal/exceptions.hpp
