+++ test/features/step_definitions/steps.hpp
@@ -8,11 +8,6 @@
 #include <iostream>
 #include <functional>
 
-#include "auth.hpp"
-#include "collection.hpp"
-#include "document.hpp"
-#include "index.hpp"
-#include "realtime.hpp"
 #include "kuzzle.hpp"
 #include "websocket.hpp"
